[SPARK-47559][SQL] Codegen Support for variant `parse_json`
### What changes were proposed in this pull request?
The PR adds Codegen Support for `parse_json`.

### Why are the changes needed?
Improve codegen coverage.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
- Add new UT.
- Pass GA.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #45714 from panbingkun/ParseJson_CodeGenerator.

Authored-by: panbingkun <panbingkun@baidu.com>
